fix/post api

+++ routes/api/post/delete.js
@@ -1,4 +1,4 @@
-const { Post, Comment } = require('../../../models');
+const { User, Post, Comment } = require('../../../models');
 const jwt = require('jsonwebtoken');
 const config = require('../../../config/index');
 const { SECRET } = config;
@@ -11,14 +11,28 @@ module.exports = async (req, res) => {
     res.status(400).json({ message: 'not token' });
   } else {
     try {
-      jwt.verify(token, SECRET);
-      await Comment.destroy({
-        where: { postId },
+      let tokenData = jwt.verify(token, SECRET);
+      let userInfo = await User.findOne({
+        attributes: ['nickname'],
+        where: { email: tokenData.email },
       });
-      await Post.destroy({
-        where: { id : postId },
+
+      let postUserInfo = await Post.findOne({
+        attributes: ['userId'],
+        where: { id: postId },
       });
-      res.status(200).json({ postDeleted: true });
+
+      if (userInfo.nickname !== postUserInfo.userId) {
+        res.status(400).json({ message: 'You are not the author of the post' });
+      } else {
+        await Comment.destroy({
+          where: { postId },
+        });
+        await Post.destroy({
+          where: { id: postId },
+        });
+        res.status(200).json({ postDeleted: true });
+      }
     } catch (err) {
       res.status(500).json({ postDeleted: false });
     }
